Title: 정수 세 개 입력받아 출력, Time: 245ms, Memory: 10MB, Status: Passed - CodeTree

+++ 231214/정수 세 개 입력받아 출력/take-three-integers-and-output.java
@@ -3,13 +3,15 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        double a = sc.nextDouble();
-        double b = sc.nextDouble();
-        double c = sc.nextDouble();
-        System.out.printf("%.3f ", a);
-        
-        System.out.printf("%.3f ", b);
-        
-        System.out.printf("%.3f", c);
+        int a = sc.nextInt();
+        int b = sc.nextInt();
+        int c = sc.nextInt();
+
+
+        System.out.printf("%d ", a);
+
+        System.out.printf("%d ", b);
+
+        System.out.printf("%d", c);
     }
 }
